Removed: fail-on-differences flag
Making the Job summary step dependent on the step's failure is wrong; we don't want the job summary if the step failed due to infrastructure problems. Rather, use the step's flag:
> if: fromJSON(steps.supersubinout.outputs.differences-found)
Combining that with the necessary failure() check (otherwise the step won't run as it defaults to success()) is cumbersome, too.
So fail-on-differences is a shortcut for the unlikely use case of just wanting a fail without any Job summary or other logging of the result, and I think it's not too much to ask to then add a minimal second step then:
,----[ step emulation of fail-on-differences ]----
| - name: Fail on differences
|   if: fromJSON(steps.supersubinout.outputs.differences-found)
|   run: exit 1
|   shell: bash
`----